Extract beforeunload listener to injectable
