+++ packages/core/src/renderer/frames/root-frame/init-root-frame.injectable.ts
@@ -9,11 +9,9 @@ import lensProtocolRouterRendererInjectable from "../../protocol-handler/lens-pr
 import catalogEntityRegistryInjectable from "../../api/catalog/entity/registry.injectable";
 import registerIpcListenersInjectable from "../../ipc/register-ipc-listeners.injectable";
 import loadExtensionsInjectable from "../load-extensions.injectable";
-import loggerInjectable from "../../../common/logger.injectable";
 import { delay } from "@k8slens/utilities";
 import { broadcastMessage } from "../../../common/ipc";
 import { bundledExtensionsLoaded } from "../../../common/ipc/extension-handling";
-import closeRendererLogFileInjectable from "../../logger/close-renderer-log-file.injectable";
 
 const initRootFrameInjectable = getInjectable({
   id: "init-root-frame",
@@ -24,10 +22,8 @@ const initRootFrameInjectable = getInjectable({
     const bindProtocolAddRouteHandlers = di.inject(bindProtocolAddRouteHandlersInjectable);
     const lensProtocolRouterRenderer = di.inject(lensProtocolRouterRendererInjectable);
     const catalogEntityRegistry = di.inject(catalogEntityRegistryInjectable);
-    const logger = di.inject(loggerInjectable);
-    const closeRendererLogFile = di.inject(closeRendererLogFileInjectable);
 
-    return async (unmountRoot: () => void) => {
+    return async () => {
       catalogEntityRegistry.init();
 
       try {
@@ -58,13 +54,6 @@ const initRootFrameInjectable = getInjectable({
       window.addEventListener("online", () => broadcastMessage("network:online"));
 
       registerIpcListeners();
-
-      window.addEventListener("beforeunload", () => {
-        logger.info("[ROOT-FRAME]: Unload app");
-
-        closeRendererLogFile();
-        unmountRoot();
-      });
     };
   },
 });
